runner/requirements.txt
fastapi
uvicorn[standard]
kafka-python
opencv-python   
requests